Add files via upload
This is a basic flutter app that works on both IOS, and Android. It has the ability to connect to an online database to add and remove information in a SQL database. A flutter app can be generated in the command line and the included dart code can be placed inside the /lib folder inside a default flutter project.